ts refactor done
