+++ .gitignore
@@ -6,6 +6,7 @@
 
 # misc
 .DS_Store
+/dist
 
 # debug
 npm-debug.log*
