+++ src/config/gpt/index.ts
@@ -1,16 +1,6 @@
 import { Configuration, OpenAIApi } from 'openai';
 import { interLeave } from '../../utils/interLeave';
-
-enum MessageRole{ 
-    System = 'system',
-    Assistant = 'assistant',
-    User = 'user',
-}
-
-type Prompt = {
-    role: MessageRole;
-    content: string;
-};
+import { MessageRole, Prompt, UserPrompt, UserRes } from '../../utils/interface';
 
 const configuration = new Configuration({
     apiKey: process.env.OPENAI_SK,
@@ -32,11 +22,8 @@ const resetInitPrompt = async () => {
 };
 
 const openai = new OpenAIApi(configuration);
-interface UserPrompt{
-    role: string;
-    content: string;
-}
-export const chatCompletion = async (content: string, prompts?: UserPrompt[], responses?: any) : Promise<any> => { // change any to Prompt[] when i figure that out
+
+export const chatCompletion = async (content: string, prompts?: UserPrompt[], responses?: UserRes[]) : Promise<any> => {
     try { 
         
         if (prompts && responses){ //if prompts exist, add them to initPrompt
