+++ src/discord/index.ts
@@ -25,8 +25,7 @@ const clientStart = async () => {
 client.on('messageCreate', async (msg) => {
     try {
         // console.log(msg);
-        //|| !msg?.author.bot && msg.channelId == '801086256711860335'
-        if (!msg?.author.bot) {
+        if (!msg?.author.bot || !msg?.author.bot && msg.channelId == '801086256711860335') {
             let msgContent = msg.content.trim();
             
             switch(true) {
