+++ src/utils/interface/index.ts
@@ -1,64 +1,77 @@
-import { Document, Model } from 'mongoose';
 
-// Define the structure of a content document
-export interface ContentDocument { 
-    global_name: string;
-    message: string;
-    message_id: string;
-    created_timestamp: number;
-    gpt_response: string;
-    tokens: {
-        prompt: number;
-        completion: number;
-        total: number;
-    }[];
-}
-
-// Define the structure of a user document
-export interface UserDocument extends Document { 
-    user_id: string;
-    global_name: string;
-    username: string;
-    content: ContentDocument[];
-    created_at: Date;
-}
-
-// Define the User model type (you might need to adjust the model name)
-export type UserModel = Model<UserDocument>; 
-
-// Define the type for the existingUser function
-export type ExistingUserFunction = ( 
-    msg: Message,
-    msgContent: string,
-    userData: UserDocument
-) => Promise<void>;
-
-
-// Define the Message type (from your previous definition)
-export interface Message { 
-    author: {
-        id: string;
-        username: string;
-        user_id: string;
-        content: string;
-        createdTimestamp: number;
-    };
-    channel:{
-        type: number;
-    },
-    createdTimestamp: number;
-    role: MessageRole;
-    content: string;
-}
-
-// Define the MessageRole enum (from your previous definition)
-export enum MessageRole{ 
+////////////////GPT/////////////////////
+enum MessageRole{ // used for gpt config
     System = 'system',
     Assistant = 'assistant',
     User = 'user',
 }
 
-export type Prompt = {
+type Prompt = { // used for gpt config
     role: MessageRole;
     content: string;
 };
+
+interface UserPrompt{ // used for user prompts going into gpt
+    role: string;
+    content: string;
+}
+
+interface UserRes{ // used for user responses going into gpt
+    role: string;
+    content: string | null;
+}
+
+export { MessageRole, Prompt, UserPrompt, UserRes };
+//////////////////////////////////////////
+
+//////////////////DirectMessages/////////////////////
+interface Token {
+    prompt: number;
+    completion: number;
+    total: number;
+}
+
+interface Content {
+    global_name: string;
+    message: string;
+    message_id: string;
+    created_timestamp: number;
+    gpt_response: string | null;
+    tokens: Token[];
+}
+
+interface UserProfile { // used for direct mesasge for users
+    user_id: string;
+    username: string;
+    content: Content[];
+    created_at: Date;
+}
+
+export { Token, Content, UserProfile };
+//////////////////////////////////////////////////////
+
+//////////////////Guilds/////////////////////
+
+interface Author {
+    user_id: string;
+    username: string;
+    global_name: string;
+    message: string;
+    message_id: string;
+    created_timestamp: number;
+}
+
+interface GuildContent {
+    author: Author[];
+    gpt_response: string | null;
+    tokens: Token[];
+}
+
+interface Guild { // used for guilds/servers
+    guild_id: string;
+    content: GuildContent[];
+    created_at: Date;
+}
+
+export { Author, GuildContent, Guild };
+//////////////////////////////////////////////////////
